+++ app/src/main/java/com/kanhaji/upastithi/screen/home/components/WeekName.kt
@@ -1,0 +1,38 @@
+package com.kanhaji.upastithi.screen.home.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import kotlinx.datetime.DayOfWeek
+import kotlin.collections.forEach
+
+@Composable
+fun DaysOfWeekTitle(daysOfWeek: List<DayOfWeek>) {
+    Row(modifier = Modifier.fillMaxWidth()) {
+        daysOfWeek.forEach { dow ->
+            Box(
+                modifier = Modifier
+                    .weight(1f)
+                    .padding(vertical = 16.dp),
+                contentAlignment = Alignment.Center
+            ) {
+                Text(
+                    modifier = Modifier.fillMaxWidth(),
+                    text = dow.name.take(3), style = MaterialTheme.typography.labelMedium,
+                    textAlign = TextAlign.Center,
+                    fontSize = 12.sp
+                )
+            }
+        }
+    }
+}
+
